refactor: Adopted DDD Structure
each domain has it's own components
controller, service, repository
