+++ src/main/java/com/lazyworking/sagupalgu/category/repository/CategoryRepository.java
@@ -1,10 +1,7 @@
-package com.lazyworking.sagupalgu.repository;
+package com.lazyworking.sagupalgu.category.repository;
 
-import com.lazyworking.sagupalgu.domain.Category;
-import com.lazyworking.sagupalgu.domain.UsedItem;
+import com.lazyworking.sagupalgu.category.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 
 public interface CategoryRepository extends JpaRepository<Category,Long> {
     Category findByName(String name);
