+++ src/main/java/com/lazyworking/sagupalgu/item/controller/UsedItemController.java
@@ -1,6 +1,6 @@
-package com.lazyworking.sagupalgu.controller;
+package com.lazyworking.sagupalgu.item.controller;
 
-import com.lazyworking.sagupalgu.service.UsedItemService;
+import com.lazyworking.sagupalgu.item.service.UsedItemService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
