+++ src/main/java/com/lazyworking/sagupalgu/item/repository/UsedItemRepository.java
@@ -1,6 +1,6 @@
-package com.lazyworking.sagupalgu.repository;
+package com.lazyworking.sagupalgu.item.repository;
 
-import com.lazyworking.sagupalgu.domain.UsedItem;
+import com.lazyworking.sagupalgu.item.domain.UsedItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
